requirements.txt
aiohttp==3.8.1
bitsandbytes==0.41.1
bottleneck==1.3.5
brotli==1.0.9
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.0
chardet==5.1.0
contourpy==1.1.1
cycler==0.12.0
debugpy==1.6.7
diffusers==0.12.1
fonttools==4.43.0
frozenlist==1.3.3
ftfy==6.1.1
jupyter-core==5.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
miniball==1.2.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.14
numexpr==2.8.4
numpy==1.24.3
opencv-python==4.7.0.68
pandas==1.5.3
pillow==9.4.0
pot==0.9.1
protobuf==3.20.3
psutil==5.9.0
pyarrow==11.0.0
pyparsing==3.1.1
pyyaml==6.0
pyzmq==25.1.0
regex==2023.6.3
responses==0.18
scikit-learn==1.3.1
scipy==1.11.3
setuptools==67.8.0
smallestenclosingcircle==0.0.1
threadpoolctl==3.2.0
tokenizers==0.11.4
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.1
urllib3==2.0.3
wheel==0.38.4
xxhash==2.0.2
yarl==1.7.2